Feat: Add global trigger for throwing behaviour and add global pKey setter
chore: update dependencies
docs: update docs for new version and some corrections
tests: update tests to check changes

BREAKING CHANGE:
__Throwable__
The throwable behaviour is customizable per method or can be overriden with a global flag.
For data fetching methods, the default behaviour is to return undefined. For a few sensible methods like empty, the default behaviour is to raise a TypeError exception.

__Parent key__
It is now possible to globally set a parent key to work on a subset without providing it as parameter each time